% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cm.R
\name{get_cm}
\alias{get_cm}
\title{Get central midfielders scouting data}
\usage{
get_cm()
}
\value{
A data frame with midfielders performance metrics.
}
\description{
Returns forward scouting metrics scaled to Football Manager–style
ratings (1–20) for easier interpretation.
}
