% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fw.R
\name{get_fw}
\alias{get_fw}
\title{Get forwards scouting data}
\usage{
get_fw()
}
\value{
A data frame with goalkeeper performance metrics.
}
\description{
Returns forward scouting metrics scaled to Football Manager–style
ratings (1–20) for easier interpretation.
}
