% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_w.R
\name{get_w}
\alias{get_w}
\title{Get wingers scouting data}
\usage{
get_w()
}
\value{
A data frame with goalkeeper performance metrics.
}
\description{
Returns wingers scouting metrics scaled to Football Manager–style
ratings (1–20) for easier interpretation.
}
