% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_wb.R
\name{get_wb}
\alias{get_wb}
\title{Get wing backs scouting data}
\usage{
get_wb()
}
\value{
A data frame with goalkeeper performance metrics.
}
\description{
Returns wing backs scouting metrics scaled to Football Manager–style
ratings (1–20) for easier interpretation.
}
