% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gk}
\alias{gk}
\title{Final Goalkeeper Ratings}
\format{
A data frame with X rows and Y variables:
\describe{
\item{Player}{Name of the goalkeeper}
\item{Final_Rating}{The 1-20 performance rating}

}
}
\usage{
gk
}
\description{
A dataset containing weighted ratings for goalkeepers across 9 leagues.
}
\keyword{datasets}
